Register global re-frame interceptor
As a side note, we have to be aware that it's easy to add default interceptors,
but as the app grows what was once thought to be *always* useful could end up
being a burden. In this particular case, I believe validating the app state is
always good during development, so there's no harm in registering it as a global
interceptor.