Remove side-effects from event handler
This commit introduces a new re-frame effect to help eliminate side-effectful
code from event handlers. It's a simple wrapper around promises. Nothing fancy
here.

There's a very similar library called re-promise[1], but it does more than I
need and honestly, it has features that are not well thought out. For example,
it has an effect :promise-n to sequentially process multiple promises, but it
does so using doseq, which is incorrect in my opinion because we can't really
loop over asynchronous code using synchronous constructs such as this.

[1] https://github.com/smogg/re-promise