Handle Etherscan rate-limit errors
Again, this commit leverages the :promise effect to move logic out of event
handlers. The result is a function in the domain.etherscan namespace that is
isolated from re-frame complexities, it's easier to test and which can be more
easily extended.

Although the "http-xhrio" effect is just data, you cannot easily abstract away
rate-limit errors, as Etherscan replies with status code 200. That means the
on-success event handler would need be complected with error handling code,
which smells bad.