Spec re-frame effect handlers using Malli
This is such an important commit to improve DX. By leveraging Malli's latest
version we can use malli.core/-instrument in ClojureScript and instrument
anonymous functions, such as the ones used by re-frame (reg-fx, etc).

In this particular commit I focused on writing a function spec for two very
simple re-frame effects (:acme.web.effect/focus-element and
:acme.web.effect/route). The major challenge was that, by default, Malli's
instrumentation uses malli.core/-fail! as a reporter and I wanted a humanized
message (the reporter throws a very difficult to read exception in the Developer
Tools).The custom reporter implemented in this commit can be further improved,
but I'm going to do it on-demand. For now, it works perfectly fine.

To make it easier to write function specs for named functions, this commit also
brings in the dependency crypticbutter.snoop, that offers a nice >defn macro and
complete support for Malli's schema representation.